Add bar chart to dashboard. Change decimal to float

+++ db/migrate/20161118234152_create_transaction.rb
@@ -2,8 +2,8 @@ class CreateTransaction < ActiveRecord::Migration
   def up
   	create_table :transactions do |t|
                 t.date :date
-                t.decimal :amount
-  		          t.text :description
+                t.float :amount
+  		t.text :description
                 t.belongs_to :user, index: true
                 t.timestamps
   	end
